parser/requirements.txt
openpyxl
xlrd
pandas
python-dotenv
